# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Readiness Analyzer API for Shopify stores
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List all connected stores
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: connectStore
      tags: [stores]
      summary: Connect a Shopify store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectStoreBody"
      responses:
        "201":
          description: Store connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          description: Invalid store URL or credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{storeId}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get a store by ID
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Disconnect a store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /stores/{storeId}/activity:
    get:
      operationId: getStoreActivity
      tags: [stores]
      summary: Get recent analysis and fix activity for a store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    Store:
      type: object
      properties:
        id:
          type: string
        domain:
          type: string
        name:
          type: string
        accessToken:
          type: string
        status:
          type: string
          enum: [connected, analyzing, analyzed, error]
        lastAnalyzed:
          type: ["string", "null"]
        overallScore:
          type: ["number", "null"]
        productCount:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - domain
        - name
        - accessToken
        - status
        - createdAt
    ConnectStoreBody:
      type: object
      properties:
        domain:
          type: string
          description: Shopify store domain (e.g., mystore.myshopify.com)
        accessToken:
          type: string
          description: Shopify Admin API access token
        name:
          type: string
          description: Display name for the store
      required:
        - domain
        - accessToken
        - name
...
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [analysis_started, analysis_completed, fix_applied, store_connected, gap_detected]
        message:
          type: string
        timestamp:
          type: string
        metadata:
          type: object
          additionalProperties: true
      required:
        - id
        - type
        - message
        - timestamp